Surah SurahManagement has not been implemented but main is partially updated. main change is in studentdetails file
